+++ src/flickr_api/__init__.py
@@ -13,7 +13,7 @@ from .exceptions import (
 )
 
 
-__version__ = "3.1"
+__version__ = "3.2"
 
 
 __all__ = [